+++ src/main/java/camp/nextstep/edu/kitchenpos/menu/application/MenuService.java
@@ -6,7 +6,7 @@ import camp.nextstep.edu.kitchenpos.menu.dao.MenuProductDao;
 import camp.nextstep.edu.kitchenpos.product.dao.ProductDao;
 import camp.nextstep.edu.kitchenpos.menu.domain.Menu;
 import camp.nextstep.edu.kitchenpos.menu.domain.MenuProduct;
-import camp.nextstep.edu.kitchenpos.product.model.Product;
+import camp.nextstep.edu.kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
