+++ src/test/java/camp/nextstep/edu/kitchenpos/product/application/ProductServiceTest.java
@@ -1,9 +1,8 @@
-package camp.nextstep.edu.kitchenpos.product.bo;
+package camp.nextstep.edu.kitchenpos.product.application;
 
 import camp.nextstep.edu.kitchenpos.product.dao.InMemoryProductDao;
 import camp.nextstep.edu.kitchenpos.product.dao.ProductDao;
-import camp.nextstep.edu.kitchenpos.product.model.Product;
-import camp.nextstep.edu.kitchenpos.product.bo.ProductBo;
+import camp.nextstep.edu.kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,16 +18,16 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("상품 Business Object 테스트 클래스")
 @ExtendWith(MockitoExtension.class)
-class ProductBoTest {
+class ProductServiceTest {
     private Product product;
 
     private ProductDao productDao = new InMemoryProductDao();
 
-    private ProductBo productBo;
+    private ProductService productService;
 
     @BeforeEach
     void setup() {
-        productBo = new ProductBo(productDao);
+        productService = new ProductService(productDao);
         product = new Product();
         product.setId(1L);
         product.setName("후라이드 치킨");
@@ -57,7 +56,7 @@ class ProductBoTest {
 
         // when then
         assertThatExceptionOfType(IllegalArgumentException.class)
-                .isThrownBy(() -> productBo.create(product));
+                .isThrownBy(() -> productService.create(product));
     }
 
     @DisplayName("[상품 생성] 상품의 금액은이 0원 미만일 경우 예외를 발생한다.")
@@ -69,14 +68,14 @@ class ProductBoTest {
 
         // when then
         assertThatExceptionOfType(IllegalArgumentException.class)
-                .isThrownBy(() -> productBo.create(product));
+                .isThrownBy(() -> productService.create(product));
     }
 
     @DisplayName("[상품 생성] 상품을 생성할 수 있다.")
     @Test
     void create() {
         // when
-        Product actual = productBo.create(product);
+        Product actual = productService.create(product);
 
         // then
         assertAll(
@@ -96,11 +95,11 @@ class ProductBoTest {
         ohterProduct.setName("양념 치킨");
         ohterProduct.setPrice(BigDecimal.valueOf(20000L));
 
-        productBo.create(product);
-        productBo.create(ohterProduct);
+        productService.create(product);
+        productService.create(ohterProduct);
 
         // when
-        List<Product> actual = productBo.list();
+        List<Product> actual = productService.list();
 
         // then
         assertThat(actual).containsExactlyInAnyOrder(product, ohterProduct);
